+++ packages/database/src/Builder/QueryBuilders/QueryBuilder.php
@@ -3,6 +3,7 @@
 namespace Tempest\Database\Builder\QueryBuilders;
 
 use Tempest\Database\Exceptions\ModelDidNotHavePrimaryColumn;
+use Tempest\Database\OnDatabase;
 use Tempest\Database\PrimaryKey;
 use Tempest\Mapper\SerializerFactory;
 
@@ -12,16 +13,14 @@ use function Tempest\Database\query;
 use function Tempest\Mapper\make;
 use function Tempest\Support\arr;
 
-/**
- * @template TModel of object
- */
-final readonly class QueryBuilder
+/** @template TModel */
+final class QueryBuilder
 {
-    /**
-     * @param class-string<TModel>|string|TModel $model
-     */
+    use OnDatabase;
+
+    /** @param class-string<TModel>|TModel|string $model */
     public function __construct(
-        private string|object $model,
+        private readonly string|object $model,
     ) {}
 
     /**
@@ -41,7 +40,7 @@ final readonly class QueryBuilder
         return new SelectQueryBuilder(
             model: $this->model,
             fields: $columns !== [] ? arr($columns)->unique() : null,
-        );
+        )->onDatabase($this->onDatabase);
     }
 
     /**
@@ -66,7 +65,7 @@ final readonly class QueryBuilder
             model: $this->model,
             rows: $values,
             serializerFactory: get(SerializerFactory::class),
-        );
+        )->onDatabase($this->onDatabase);
     }
 
     /**
@@ -88,7 +87,7 @@ final readonly class QueryBuilder
             model: $this->model,
             values: $values,
             serializerFactory: get(SerializerFactory::class),
-        );
+        )->onDatabase($this->onDatabase);
     }
 
     /**
@@ -106,7 +105,7 @@ final readonly class QueryBuilder
      */
     public function delete(): DeleteQueryBuilder
     {
-        return new DeleteQueryBuilder($this->model);
+        return new DeleteQueryBuilder($this->model)->onDatabase($this->onDatabase);
     }
 
     /**
@@ -124,7 +123,7 @@ final readonly class QueryBuilder
         return new CountQueryBuilder(
             model: $this->model,
             column: $column,
-        );
+        )->onDatabase($this->onDatabase);
     }
 
     /**
@@ -135,7 +134,7 @@ final readonly class QueryBuilder
      * query(User::class)->new(name: 'Frieren');
      * ```
      *
-     * @return TModel
+     * @return TModel|object<TModel>
      */
     public function new(mixed ...$params): object
     {
@@ -150,7 +149,7 @@ final readonly class QueryBuilder
      * query(User::class)->findById(1);
      * ```
      *
-     * @return TModel
+     * @return TModel|object<TModel>|null
      */
     public function findById(string|int|PrimaryKey $id): ?object
     {
@@ -169,9 +168,9 @@ final readonly class QueryBuilder
      * query(User::class)->resolve(1);
      * ```
      *
-     * @return TModel
+     * @return TModel|object<TModel>|null
      */
-    public function resolve(string|int|PrimaryKey $id): object
+    public function resolve(string|int|PrimaryKey $id): ?object
     {
         if (! inspect($this->model)->hasPrimaryKey()) {
             throw ModelDidNotHavePrimaryColumn::neededForMethod($this->model, 'resolve');
@@ -188,7 +187,7 @@ final readonly class QueryBuilder
      * query(User::class)->get(1);
      * ```
      *
-     * @return TModel|null
+     * @return TModel|object<TModel>|null
      */
     public function get(string|int|PrimaryKey $id, array $relations = []): ?object
     {
@@ -247,7 +246,7 @@ final readonly class QueryBuilder
      * query(User::class)->create(name: 'Frieren', kind: Kind::ELF);
      * ```
      *
-     * @return TModel
+     * @return TModel|object<TModel>
      */
     public function create(mixed ...$params): object
     {
@@ -255,9 +254,7 @@ final readonly class QueryBuilder
 
         $model = $this->new(...$params);
 
-        $id = query($this->model)
-            ->insert($model)
-            ->execute();
+        $id = $this->insert($model)->execute();
 
         $inspector = inspect($this->model);
         $primaryKeyProperty = $inspector->getPrimaryKeyProperty();
@@ -286,7 +283,7 @@ final readonly class QueryBuilder
      *
      * @param array<string,mixed> $find Properties to search for in the existing model.
      * @param array<string,mixed> $update Properties to update or set on the model if it is found or created.
-     * @return TModel
+     * @return TModel|object<TModel>
      */
     public function findOrNew(array $find, array $update): object
     {
@@ -318,7 +315,7 @@ final readonly class QueryBuilder
      *
      * @param array<string,mixed> $find Properties to search for in the existing model.
      * @param array<string,mixed> $update Properties to update or set on the model if it is found or created.
-     * @return TModel
+     * @return TModel|object<TModel>
      */
     public function updateOrCreate(array $find, array $update): object
     {
@@ -338,6 +335,7 @@ final readonly class QueryBuilder
         }
 
         query($model)
+            ->onDatabase($this->onDatabase)
             ->update(...$update)
             ->execute();